Updates to classifier tasks. Moved into pipeline-independent packages
